This object is only used once, so inline it instead
